+++ examples/poolex_example/lib/poolex_example.ex
@@ -1,0 +1,2 @@
+defmodule PoolexExample do
+end